Loan service implementation class - first push & merge

+++ src/main/java/fundMe/data/models/Loan.java
@@ -5,6 +5,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -16,8 +17,8 @@ public class Loan {
     private String loanID;
     private String borrowerId;
     private String lenderId;
-    private Long amount;
-    private int interestRate;
+    private Double amount;
+    private Double interestRate;
     private String paymentPlan;
     private Boolean isCollected;
     private Boolean isPaid;
